refactor: extract ConnectionsListPanel (#3501 Phase 1n)
